requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==3.0.6
django-extensions==2.2.9
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
django-werkzeug-debugger-runserver==0.3.1
get==2019.4.13
idna==2.9
Pillow==7.1.2
post==2019.4.13
public==2019.4.13
pycparser==2.20
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyperclip==1.8.0
pytz==2020.1
query-string==2019.4.13
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
Werkzeug==1.0.1
